refactor(core): extrai WEAPONS para data/weapons.js
Isola a tabela de configuracao de armas (dano, recuo, fire rate) num modulo de dados puro, quebrando o ciclo de dependencia onde botbrain importava a classe de volta do jogo.

Agent: Antigravity (Gemini)
